Marginal Zero Inflated Bivariate Poisson Distribution - Laksh

This function fit a marginal Zero Inflated Bivariate Poisson distribution under the of Lakshminarayana et. al (1993).

marZIBP_Laksh(mu1.fo, mu2.fo, psi.fo, data, initial.values = NULL)

Arguments

mu1.fo

a formula object to explain the marginal mean \(\mu_1\), the first response \(X_1\) on the left of an ~ operator, and the terms, separated by operators, on the right.

mu2.fo

a formula object to explain the marginal mean \(\mu_2\), the second response \(X_2\) on the left of an ~ operator, and the terms, separated by operators, on the right.

psi.fo

a formula object to explain the \(\psi\) parameter. An example could be: ~ age + salary.

data

a data frame containing the variables occurring in the formula2.

initial.values

a vector with possible initial values for the parameters.

Value

Returns a object with class marZIBP_Laksh.

References

Lakshminarayana, J., Pandit, S. N., & Srinivasa Rao, K. (1999). On a bivariate Poisson distribution. Communications in Statistics-Theory and Methods, 28(2), 267-276.

Author

Freddy Hernandez-Barajas, fhernanb@unal.edu.co

Examples

# Example 1 ---------------------------------------------------------------
l1 <- 3
l2 <- 4
alpha <- -0.90
psi <- 0.2

data1 <- rZIBP_Laksh(n=5000, l1=l1, l2=l2, alpha=alpha, psi=psi)
data1 <- as.data.frame(data1)

# To fit the model
mod1 <- NULL
mod1 <- marZIBP_Laksh(mu1.fo=X1~1,
                      mu2.fo=X2~1,
                      psi.fo=~1,
                      data=data1)
#> N = 4, M = 5 machine precision = 2.22045e-16
#> At X0, 0 variables are exactly at the bounds
#> At iterate     0  f=        23498  |proj g|=         7887
#> At iterate    10  f =        18582  |proj g|=        1.2608
#> At iterate    20  f =        18578  |proj g|=         11.82
#> 
#> iterations 25
#> function evaluations 28
#> segments explored during Cauchy searches 26
#> BFGS updates skipped 0
#> active bounds at final generalized Cauchy point 1
#> norm of the final projected gradient 0.00795453
#> final function value 18578.3
#> 
#> F = 18578.3
#> final  value 18578.317221 
#> converged

# To obtain the usual summary table
summary(mod1)
#> ---------------------------------------------------------------
#> Fixed effects for log(mu1) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) 0.889808   0.011423  77.895 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for log(mu2) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) 1.161597   0.010533  110.28 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for logit(psi) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) -1.41911    0.03577 -39.673 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Estimation for alpha 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept) -1.27769    0.56759 -2.2511  0.02438 *
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------

# To explore the estimations of l1, l2, mu and p

# To obtain E(Y1)=v1 and E(Y2)=v2
mod1$fitted.mu1[1]
#> [1] 2.434662
mod1$fitted.mu2[1]
#> [1] 3.195032

# To compare sample means with v1 and v2
colMeans(data1)
#>     X1     X2 
#> 2.4340 3.1948 

# To obtain alpha and psi
mod1$fitted.alpha
#> [1] -1.277685
mod1$fitted.psi[1]
#> [1] 0.1948003

# To obtain l1 and l2
mod1$fitted.l1[1]
#> [1] 3.023675
mod1$fitted.l2[1]
#> [1] 3.967999

# Example 2 ---------------------------------------------------------------

gen_data_ZIBP_Laksh <- function(n=10) {
  # To generate the covariates
  x1 <- runif(n=n)
  x2 <- runif(n=n)

  # To generate the means
  mu1 <- exp(-2 + 3.5 * x1 + 2.7 * x2)
  mu2 <- exp(-1 + 1.3 * x1 + 2.1 * x2)

  # To generate the p
  logit_inv <- function(x) exp(x) / (1+exp(x))
  psi  <- logit_inv(-2.4 + 1.2 * x2)

  alpha <- -1

  # To obtain lambdas
  l1 <- mu1 / (1-psi)
  l2 <- mu2 / (1-psi)

  # To generate Y1 and Y2
  y <- NULL
  for (i in 1:n)
    y <- rbind(y, rZIBP_Laksh(n=1, l1=l1[i], l2=l2[i],
                              alpha=alpha, psi=psi[i]))

  # To create the dataset
  dataset <- data.frame(y1=y[,1], y2=y[,2],
                        x1=x1, x2=x2,
                        mu1=mu1, mu2=mu2,
                        alpha=alpha, psi=psi,
                        l1=l1, l2=l2)

  return(dataset)
}

data2 <- gen_data_ZIBP_Laksh(n=100)
head(data2, n=8)
#>   y1 y2        x1         x2        mu1       mu2 alpha        psi         l1
#> 1  0  0 0.8687742 0.84500335 27.7217941 6.7121113    -1 0.20004774 34.6543108
#> 2  5  2 0.9243132 0.08306297  4.3032307 1.4565044    -1 0.09109609  4.7345277
#> 3  7  1 0.6553841 0.61587940  7.0760692 3.1435289    -1 0.15963522  8.4202354
#> 4  0  1 0.2126957 0.38297301  0.8012856 1.0840974    -1 0.12560036  0.9163838
#> 5  0  2 0.1624281 0.38343986  0.6728626 1.0165151    -1 0.12566190  0.7695680
#> 6  6  8 0.4733979 0.76858729  5.6523888 3.4193503    -1 0.18577579  6.9420544
#> 7  1  1 0.3076361 0.20545598  0.6917417 0.8448336    -1 0.10400901  0.7720409
#> 8  0  2 0.1824082 0.35888781  0.6753161 0.9908373    -1 0.12246037  0.7695563
#>          l2
#> 1 8.3906399
#> 2 1.6024845
#> 3 3.7406718
#> 4 1.2398192
#> 5 1.1626110
#> 6 4.1995193
#> 7 0.9429042
#> 8 1.1291083

mod2 <- NULL
mod2 <- marZIBP_Laksh(mu1.fo=y1~x1+x2,
                      mu2.fo=y2~x1+x2,
                      psi.fo=~x1+x2,
                      data=data2)
#> N = 10, M = 5 machine precision = 2.22045e-16
#> At X0, 0 variables are exactly at the bounds
#> At iterate     0  f=       991.65  |proj g|=       368.32
#> At iterate    10  f =       352.83  |proj g|=        9.7335
#> At iterate    20  f =          349  |proj g|=        1.8343
#> At iterate    30  f =       348.35  |proj g|=        4.3344
#> At iterate    40  f =       347.07  |proj g|=        1.6284
#> At iterate    50  f =       346.98  |proj g|=       0.52719
#> At iterate    60  f =       346.86  |proj g|=        1.4468
#> At iterate    70  f =       346.81  |proj g|=       0.12075
#> At iterate    80  f =       346.81  |proj g|=       0.17784
#> At iterate    90  f =       346.81  |proj g|=       0.12679
#> At iterate   100  f =       346.81  |proj g|=      0.048033
#> At iterate   110  f =       346.81  |proj g|=      0.019848
#> 
#> iterations 110
#> function evaluations 123
#> segments explored during Cauchy searches 111
#> BFGS updates skipped 0
#> active bounds at final generalized Cauchy point 1
#> norm of the final projected gradient 0.0198484
#> final function value 346.807
#> 
#> F = 346.807
#> final  value 346.807262 
#> converged

summary(mod2)
#> ---------------------------------------------------------------
#> Fixed effects for log(mu1) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) -1.88753    0.21490 -8.7833 < 2.2e-16 ***
#> x1           3.50727    0.22937 15.2912 < 2.2e-16 ***
#> x2           2.58496    0.22816 11.3297 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for log(mu2) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) -1.33797    0.25573 -5.2319 1.678e-07 ***
#> x1           1.36827    0.27690  4.9413 7.759e-07 ***
#> x2           2.65974    0.31045  8.5674 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for logit(psi) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept)  -2.2101     1.0672 -2.0709  0.03837 *
#> x1           -1.3218     1.1359 -1.1636  0.24457  
#> x2            1.6709     1.1712  1.4267  0.15367  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Estimation for alpha 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value Pr(>|z|)
#> (Intercept)  -1.1940     1.5128 -0.7892     0.43
#> ---------------------------------------------------------------

mod2$fitted.alpha
#> [1] -1.193957