Marginal Zero Inflated Bivariate Poisson Distribution - Laksh

This function fit a marginal Zero Inflated Bivariate Poisson distribution under the of Lakshminarayana et. al (1993).

marZIBP_Laksh(mu1.fo, mu2.fo, psi.fo, data, initial.values = NULL)

Arguments

mu1.fo

a formula object to explain the marginal mean \(\mu_1\), the first response \(X_1\) on the left of an ~ operator, and the terms, separated by operators, on the right.

mu2.fo

a formula object to explain the marginal mean \(\mu_2\), the second response \(X_2\) on the left of an ~ operator, and the terms, separated by operators, on the right.

psi.fo

a formula object to explain the \(\psi\) parameter. An example could be: ~ age + salary.

data

a data frame containing the variables occurring in the formula2.

initial.values

a vector with possible initial values for the parameters.

Value

Returns a object with class "marZIBPLaksh".

References

Lakshminarayana, J., Pandit, S. N., & Srinivasa Rao, K. (1999). On a bivariate Poisson distribution. Communications in Statistics-Theory and Methods, 28(2), 267-276.

Author

Freddy Hernandez-Barajas, fhernanb@unal.edu.co

Examples

# Example 1 ---------------------------------------------------------------
l1 <- 3
l2 <- 4
alpha <- -0.90
psi <- 0.2

data1 <- rZIBP_Laksh(n=100, l1=l1, l2=l2, alpha=alpha, psi=psi)
data1 <- as.data.frame(data1)

# To fit the model
mod1 <- NULL
mod1 <- marZIBP_Laksh(mu1.fo=X1~1,
                      mu2.fo=X2~1,
                      psi.fo=~1,
                      data=data1)
#> N = 4, M = 5 machine precision = 2.22045e-16
#> At X0, 0 variables are exactly at the bounds
#> At iterate     0  f=       457.65  |proj g|=       118.24
#> At iterate    10  f =       374.48  |proj g|=       0.25859
#> At iterate    20  f =       374.19  |proj g|=     4.101e-06
#> 
#> iterations 20
#> function evaluations 25
#> segments explored during Cauchy searches 21
#> BFGS updates skipped 0
#> active bounds at final generalized Cauchy point 1
#> norm of the final projected gradient 4.10097e-06
#> final function value 374.19
#> 
#> F = 374.19
#> final  value 374.190479 
#> converged

# To obtain the usual summary table
summary(mod1)
#> ---------------------------------------------------------------
#> Fixed effects for log(mu1) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) 0.967292   0.080398  12.031 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for log(mu2) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) 1.018234   0.079236  12.851 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for logit(psi) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) -1.32493    0.24659 -5.3729 7.747e-08 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Estimation for alpha 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value Pr(>|z|)
#> (Intercept)  -1.2782     5.7430 -0.2226   0.8239
#> ---------------------------------------------------------------

# To explore the estimations of l1, l2, mu and p

# To obtain E(Y1)=v1 and E(Y2)=v2
mod1$fitted.mu1[1]
#> [1] 2.63081
mod1$fitted.mu2[1]
#> [1] 2.768301

# To compare sample means with v1 and v2
colMeans(data1)
#>   X1   X2 
#> 2.63 2.77 

# To obtain alpha and psi
mod1$fitted.alpha
#> [1] -1.278169
mod1$fitted.psi[1]
#> [1] 0.21

# To obtain l1 and l2
mod1$fitted.l1[1]
#> [1] 3.33014
mod1$fitted.l2[1]
#> [1] 3.504178

# Example 2 ---------------------------------------------------------------

gen_data_ZIBP_Laksh <- function(n=10) {
  # To generate the covariates
  x1 <- runif(n=n)
  x2 <- runif(n=n)

  # To generate the means
  mu1 <- exp(-2 + 3.5 * x1 + 2.7 * x2)
  mu2 <- exp(-1 + 1.3 * x1 + 2.1 * x2)

  # To generate the p
  logit_inv <- function(x) exp(x) / (1+exp(x))
  psi  <- logit_inv(-2.4 + 1.2 * x2)

  alpha <- -1

  # To obtain lambdas
  l1 <- mu1 / (1-psi)
  l2 <- mu2 / (1-psi)

  # To generate Y1 and Y2
  y <- NULL
  for (i in 1:n)
    y <- rbind(y, rZIBP_Laksh(n=1, l1=l1[i], l2=l2[i],
                              alpha=alpha, psi=psi[i]))

  # To create the dataset
  dataset <- data.frame(y1=y[,1], y2=y[,2],
                        x1=x1, x2=x2,
                        mu1=mu1, mu2=mu2,
                        alpha=alpha, psi=psi,
                        l1=l1, l2=l2)

  return(dataset)
}

data2 <- gen_data_ZIBP_Laksh(n=100)
head(data2, n=8)
#>   y1 y2        x1         x2        mu1       mu2 alpha        psi         l1
#> 1  0  0 0.3187409 0.05085399  0.4737379 0.6194995    -1 0.08794608  0.5194187
#> 2 27  6 0.8059563 0.80323340 19.8772775 5.6662829    -1 0.19214692 24.6050650
#> 3  0  0 0.4867537 0.58377585  3.5960381 2.3601120    -1 0.15453463  4.2533240
#> 4  2  3 0.4312844 0.58548322  2.9751670 2.2038030    -1 0.15480251  3.5200850
#> 5  1  6 0.6470496 0.32102598  3.1001288 1.6741725    -1 0.11766099  3.5135348
#> 6 28  6 0.7340863 0.95843950 23.5020778 7.1494483    -1 0.22272219 30.2363937
#> 7  9  2 0.8765104 0.25387116  5.7732700 1.9593112    -1 0.10954915  6.4835358
#> 8  4  2 0.7769337 0.22157274  3.7341276 1.6085232    -1 0.10582519  4.1760600
#>          l2
#> 1 0.6792356
#> 2 7.0140018
#> 3 2.7914946
#> 4 2.6074415
#> 5 1.8974254
#> 6 9.1980605
#> 7 2.2003587
#> 8 1.7988913

mod2 <- NULL
mod2 <- marZIBP_Laksh(mu1.fo=y1~x1+x2,
                      mu2.fo=y2~x1+x2,
                      psi.fo=~x1+x2,
                      data=data2)
#> N = 10, M = 5 machine precision = 2.22045e-16
#> At X0, 0 variables are exactly at the bounds
#> At iterate     0  f=         1259  |proj g|=       503.06
#> At iterate    10  f =       368.75  |proj g|=        8.4306
#> At iterate    20  f =       362.02  |proj g|=        17.597
#> At iterate    30  f =       360.86  |proj g|=        4.4996
#> At iterate    40  f =       359.36  |proj g|=       0.63549
#> At iterate    50  f =       359.33  |proj g|=       0.60369
#> At iterate    60  f =       359.31  |proj g|=         0.247
#> At iterate    70  f =       359.29  |proj g|=       0.23631
#> At iterate    80  f =       359.28  |proj g|=       0.16461
#> At iterate    90  f =       359.28  |proj g|=      0.049611
#> At iterate   100  f =       359.28  |proj g|=       0.21538
#> At iterate   110  f =       359.28  |proj g|=      0.079214
#> At iterate   120  f =       359.28  |proj g|=      0.025682
#> At iterate   130  f =       359.28  |proj g|=      0.025955
#> 
#> iterations 133
#> function evaluations 149
#> segments explored during Cauchy searches 134
#> BFGS updates skipped 0
#> active bounds at final generalized Cauchy point 0
#> norm of the final projected gradient 0.00904005
#> final function value 359.277
#> 
#> F = 359.277
#> final  value 359.277236 
#> converged

summary(mod2)
#> ---------------------------------------------------------------
#> Fixed effects for log(mu1) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) -2.12853    0.27868 -7.6380 2.207e-14 ***
#> x1           3.78062    0.28187 13.4125 < 2.2e-16 ***
#> x2           2.55565    0.30753  8.3102 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for log(mu2) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value  Pr(>|z|)    
#> (Intercept) -1.00729    0.28951 -3.4793 0.0005027 ***
#> x1           1.49939    0.31057  4.8279 1.380e-06 ***
#> x2           1.97153    0.34494  5.7156 1.093e-08 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Fixed effects for logit(psi) 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept) -1.66833    1.01199 -1.6486  0.09924 .
#> x1          -0.79232    0.90031 -0.8801  0.37883  
#> x2           1.42272    1.27218  1.1183  0.26343  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> ---------------------------------------------------------------
#> Estimation for alpha 
#> ---------------------------------------------------------------
#>             Estimate Std. Error z value Pr(>|z|)
#> (Intercept)  0.19167    1.93994  0.0988   0.9213
#> ---------------------------------------------------------------

mod2$fitted.alpha
#> [1] 0.1916694